requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
dj-database-url==0.4.2
Django==1.11.7
django-extensions==1.9.8
-e git+http://github.com/Verbozeteam/django-webpack-loader.git@master#egg=django_webpack_loader
djangorestframework==3.7.3
gunicorn==19.7.1
hyperlink==17.3.1
incremental==17.5.0
msgpack-python==0.4.8
psycopg2==2.7.3.1
pytz==2017.3
raven==6.9.0
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
typing==3.6.2
whitenoise==3.3.1
zope.interface==4.4.3
